+++ tests/h.GitApi.js
